refactor(app): move shared bundle helpers out of shell
